refactored by mainly changing module structure
renamed symmetry_function/ -> dataset/
  separated symmetry_function/utils.py
    -> dataset/scatter_dataset.py and dataset/generator.py
  separated symmetry_function/file_io.py
    -> dataset/construct_datasets.py and format/vasp.py,xyz.py
  merged symmetry_function/dataset.py,symmetry_functions.py
    -> dataset/symmetry_function_dataset.py
moved `parse_xxx()` methods -> format/
renamed util.py -> utils.py
changed all `Path` object as absolute path
created __init__.py in subpackages
