+++ hdnnpy/__init__.py
@@ -1,1 +1,1 @@
-# coding=utf-8
+# -*- coding: utf-8 -*-
