+++ hdnnpy/format/__init__.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+
+from .vasp import parse_poscars
+from .xyz import parse_xyz
+
+__all__ = ['parse_poscars', 'parse_xyz']
